Dask related fixes and optimizations (#461)
* Protocol was the problem, switched to inproc for process=False

* Taking care of garbage collection for delayed objects

* Another big improvement

* slight optimization mentioning that tasks are not pure function, i.e they rely on side effects instead of being isolated

* minor mods

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Limiting worker memory to prevent hogging resources on a laptop
